fixed three (!) bugs in gillespie.
First: a scalar addition that was accidentally being done to a whole
array, when updating the chemical counts.

Second: an array copy of init_amts, which was being done by reference
so that init_amts got messed up by the simulated-data generation. This
may have been causing the high rejection rates in first rounds of all
my tests.

Third: Another shallow array copy was affecting the x_path contents in
the data generation.